Replace SpriteAnimBody with AnimationBehavior
