+++ src/TilemapScene.cpp
@@ -1,20 +1,20 @@
 #include "Scenes/Scene.h"
 #include "Scripts/Script.h"
 #include "World/WorldInstance.h"
-#include "Body/SpriteAnimBody.h"
 #include "Scripts/CommonScripts.h"
+#include "Behavior/AnimationBehavior.h"
 
 namespace CGEngine {
 	class TilemapScene : public Scene {
     public:
         TilemapScene() { loadEvent = new Script(mainEvt); displayName = "TilemapScene"; }
 
-        SpriteAnimBody* player = nullptr;
         id_t gridId = 0U;
         vector<int> tileTypes = { 0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26,27,28,29,30,31,32,33,34,35,36,37,38,39,40,41,42,43,44,45,46,47,48,49,50,51,52,53,54 };
         int assignedType = 0;
         optional<Vector2f> clickedTilemapPos = nullopt;
         string tilemapDataPath = "mapData_.txt";
+        Behavior* beh = new AnimationBehavior(nullptr, "ts", { 32,32 });
 
         ScriptEvent saveTilemapDataEvt = [](ScArgs args) { args.caller->get<Tilemap*>()->saveMapData(); };
         ScriptEvent tileClickEvt = [this](ScArgs args) {
@@ -90,25 +90,25 @@ namespace CGEngine {
                     args.caller->addOverlapMouseReleaseScript(new Script(toolboxTileClickEvt), Mouse::Button::Left);
                     }));
 
-                ScriptEvent clickUpdateScript = [this](ScArgs args) {
-                    MouseReleaseInput* mouseEvt = args.script->getInput().getDataPtr<MouseReleaseInput>("evt");
-                    if (mouseEvt == nullptr || clickedTilemapPos == nullopt) return;
-
-                    V2f mouseGlobalPos = screen->viewToGlobal(mouseEvt->position);
-                    Body* worldGrid = world->bodies.get(gridId);
-                    if (worldGrid->contains(mouseGlobalPos)) {
-                        SpriteAnimBody* body = new SpriteAnimBody("animation.png", AnimationParameters({ 32,32 }, 30.f), Transformation(), worldGrid);
-                        body->setTimer(0.5, new Script([](ScArgs args) { world->deleteBody(args.caller); }));
-                        Vector2f pos = (Vector2f)clickedTilemapPos.value();
-                        int d = (int)(pos.x) % 32;
-                        pos.x = pos.x - d;
-                        d = (int)(pos.y) % 32;
-                        pos.y = pos.y - d;
-                        body->setPosition(pos * 1.5f);
-                    }
-                    };
-                world->getRoot()->addMouseReleaseScript(clickUpdateScript, Mouse::Button::Left);
-                world->getRoot()->addMouseReleaseScript(clickUpdateScript, Mouse::Button::Right);
+                //ScriptEvent clickUpdateScript = [this](ScArgs args) {
+                //    MouseReleaseInput* mouseEvt = args.script->getInput().getDataPtr<MouseReleaseInput>("evt");
+                //    if (mouseEvt == nullptr || clickedTilemapPos == nullopt) return;
+                //
+                //    V2f mouseGlobalPos = screen->viewToGlobal(mouseEvt->position);
+                //    Body* worldGrid = world->bodies.get(gridId);
+                //    if (worldGrid->contains(mouseGlobalPos)) {
+                //        SpriteAnimBody* body = new SpriteAnimBody("animation.png", AnimationParameters({ 32,32 }, 30.f), Transformation(), worldGrid);
+                //        body->setTimer(0.5, new Script([](ScArgs args) { world->deleteBody(args.caller); }));
+                //        Vector2f pos = (Vector2f)clickedTilemapPos.value();
+                //        int d = (int)(pos.x) % 32;
+                //        pos.x = pos.x - d;
+                //        d = (int)(pos.y) % 32;
+                //        pos.y = pos.y - d;
+                //        body->setPosition(pos * 1.5f);
+                //    }
+                //    };
+                //world->getRoot()->addMouseReleaseScript(clickUpdateScript, Mouse::Button::Left);
+                //world->getRoot()->addMouseReleaseScript(clickUpdateScript, Mouse::Button::Right);
 
                 args.script->setOutput(map<string, any>({ {"gridId", gridId} }));
             };
@@ -138,166 +138,8 @@ namespace CGEngine {
                 player->addStartScript(keyboardMoveController);
                 player->addStartScript(keyboardRotateController);
 
-                //Declare the input (unmodified) and data (modified) parameters of the animation behavior
-                Behavior* animationBehavior = new Behavior(player);
-                animationBehavior->input = DataMap(map<string, any>({
-                    {"maxFrameRate",15.0f},
-                    {"speed",1.0f},
-                    {"maxFrame",7},
-                    {"looping",true},
-                    {"startRunning",false},
-                    {"sheetSize",{spriteTex->getSize()}}
-                }));
-                animationBehavior->data = DataMap(map<string, any>({
-                    {"rect",spriteRect},
-                    {"frame",0},
-                    {"state",AnimationState::Ready},
-                    {"frameTime",0.f},
-                    {"animationStartPos",Vector2i({0,0})}
-                }));
-
-                //When the Behavior is started (when the Body is started)
-                animationBehavior->scripts.addScript(onStartEvent, new Script([](ScArgs args) {
-                    IntRect spriteRect = args.behavior->data.getData<IntRect>("rect");
-                    bool startRunning = args.behavior->input.getData<bool>("startRunning");
-
-                    args.caller->get<Sprite*>()->setTextureRect(spriteRect);
-                    if (startRunning) {
-                        args.behavior->scripts.callDomain("startAnimation",args.behavior);
-                    }
-                }));
-
-                //Animation scripts
-                animationBehavior->scripts.addScript("calculateAnimLength", new Script([](ScArgs args) {
-                    float maxFrameRate = args.behavior->input.getData<float>("maxFrameRate");
-                    float speed = args.behavior->input.getData<float>("speed");
-                    float frameRate = maxFrameRate * abs(speed);
-                    float frameLength = 1.f / frameRate;
-                    args.behavior->data.addData("frameLength", frameLength);
-                }));
-
-                animationBehavior->scripts.addScript("pauseAnimation", new Script([](ScArgs args) {
-                    id_t animationUpdateScriptId = args.behavior->data.getData<id_t>("animUpdateId");
-                    args.behavior->data.addData("state", AnimationState::Paused);
-                    args.behavior->scripts.removeScript(onUpdateEvent, animationUpdateScriptId,true);
-                }));
-
-                animationBehavior->scripts.addScript("endAnimation", new Script([](ScArgs args) {
-                    args.behavior->data.addData("frameTime", 0.f);
-
-                    Vector2i startingPos = args.behavior->data.getData<Vector2i>("animationStartPos");
-                    IntRect spriteRect = args.behavior->data.getData<IntRect>("rect");
-                    spriteRect.position = startingPos;
-                    args.behavior->data.addData("rect", spriteRect);
-                    args.caller->get<Sprite*>()->setTextureRect(spriteRect);
-
-                    args.behavior->scripts.callDomain("pauseAnimation", args.behavior);
-                    args.behavior->data.addData("state", AnimationState::Ready);
-
-                    args.behavior->data.addData("frame", 0);
-                }));
-
-                animationBehavior->scripts.addScript("startAnimation", new Script([](ScArgs args) {
-                    args.behavior->scripts.callDomain("calculateAnimLength", args.behavior);
-                    args.behavior->data.addData("state", AnimationState::Running);
-                    
-                    id_t animationUpdateScriptId = args.behavior->scripts.addScript(onUpdateEvent, new Script([](ScArgs args) {
-                        args.behavior->scripts.callDomain("animate", args.behavior);
-                    }));
-                    args.behavior->data.addData("animUpdateId", animationUpdateScriptId);
-                }));
-
-                animationBehavior->scripts.addScript("animate", new Script([](ScArgs args) {
-                    int frame = args.behavior->data.getData<int>("frame");
-                    IntRect spriteRect = args.behavior->data.getData<IntRect>("rect");
-                    float frameLength = args.behavior->data.getData<float>("frameLength");
-                    float state = args.behavior->data.getData<AnimationState>("state");
-
-                    bool looping = args.behavior->input.getData<bool>("looping");
-                    bool startRunning = args.behavior->input.getData<bool>("startRunning");
-                    Vector2u spriteSheetSize = args.behavior->input.getData<Vector2u>("sheetSize");
-                    Vector2i animationStartPos = args.behavior->data.getData<Vector2i>("animationStartPos");
-                    int maxFrame = args.behavior->input.getData<int>("maxFrame");
-
-                    if (frameLength <= 0) return;
-                    float frameTime = args.behavior->data.getData<float>("frameTime") + time.getDeltaSec();
-                    if (frameTime > frameLength) {
-                        while ((frameTime -= frameLength) >= 0.f) {
-                            if (maxFrame > 0 && frame >= maxFrame) {
-                                if (looping) {
-                                    frame = 0;
-                                    spriteRect.position = animationStartPos;
-                                } else {
-                                    args.behavior->scripts.callDomain("pauseAnimation", args.behavior);
-                                    break;
-                                }
-                            }
-
-                            //Advance frame rect
-                            if (maxFrame < 0 || frame <= (unsigned)maxFrame) {
-                                if ((unsigned)(spriteRect.position.x + spriteRect.size.x) < spriteSheetSize.x) {
-                                    spriteRect.position.x += spriteRect.size.x;
-                                    frame++;
-                                }
-                                else {
-                                    if ((unsigned)(spriteRect.position.y + spriteRect.size.y) < spriteSheetSize.y) {
-                                        spriteRect.position.y += spriteRect.size.y;
-                                        spriteRect.position.x = 0;
-                                        frame++;
-                                    }
-                                    else {
-                                        if (looping) {
-                                            spriteRect.position = animationStartPos;
-                                            frame++;
-                                        }
-                                        else {
-                                            args.behavior->scripts.callDomain("pauseAnimation", args.behavior);
-                                            break;
-                                        }
-                                    }
-                                }
-                            }
-                        }
-                        args.caller->get<Sprite*>()->setTextureRect(spriteRect);
-                        args.behavior->data.addData("frame", frame);
-                        args.behavior->data.addData("rect", spriteRect);
-                    }
-                    args.behavior->data.addData("frameTime", frameTime);
-                }));
-
-                animationBehavior->scripts.addScript(onUpdateEvent, new Script([&playerSpeed](ScArgs args) {
-                    Vector2f facing = args.behavior->data.getData<Vector2f>("facing");
-                    Vector2f lastMove = args.behavior->data.getData<Vector2f>("lastMove");
-
-                    if (facing != lastMove && lastMove != V2f({ 0,0 })) {
-                        if (lastMove == Vector2f({ 0, 1 })) {
-                            args.behavior->data.addData("animationStartPos", Vector2i({ 0,0 }));
-                        } else if (lastMove == Vector2f({ 0, -1 })) {
-                            args.behavior->data.addData("animationStartPos", Vector2i({ 0,96 }));
-                        } else if (lastMove == Vector2f({ 1, 0 })) {
-                            args.behavior->data.addData("animationStartPos", Vector2i({ 0,32 }));
-                        } else {
-                            args.behavior->data.addData("animationStartPos", Vector2i({ 0,64 }));
-                        }
-
-                        facing = lastMove;
-                        args.behavior->data.addData("facing", facing);
-                        args.behavior->scripts.callDomain("endAnimation", args.behavior);
-                    }
-                    lastMove = { 0,0 };
-                    args.behavior->data.addData("lastMove", lastMove);
-                }));
-
-                animationBehavior->scripts.addScript("OnTranslate", new Script([this](ScArgs args) {
-                    AnimationState animState = args.behavior->data.getData<AnimationState>("state");
-                    if (animState != AnimationState::Running) {
-                        args.behavior->scripts.callDomain("startAnimation", args.behavior);
-                    }
-                    Vector2f lastMove = args.script->getInput().getData<Vector2f>("evt");
-                    args.behavior->data.addData("lastMove", lastMove);
-                }));
-
-                //Finally, add the animationBehavior to the Body and get the id
+                Behavior* animationBehavior = new AnimationBehavior(player, "triceratops.png", { 32,32 });
+                //Add the animationBehavior to the Body and get the id
                 id_t animBehaviorId = player->addBehavior(animationBehavior);
 
                 //Add KeyRelease scripts that call the "endAnimation" domain on the animationBehavior via its id
